requirements.txt

django
python-dotenv
djangorestframework
mysqlclient
django-userforeignkey
markdown
django-filter
pip
#install
django-disposable-email-checker
django-extensions
six
django-q
dj-stripe
stripe
pillow
django-tinymce